Remove CachedBitcoind from BitcoinSWalletTest
